fix edge 18 incompatibiltiy in url-regex lib -- #3850
