+++ src/smc-util-node/package.json
@@ -20,6 +20,7 @@
     "shell-escape": "^0.2.0",
     "sqlite3": "^4.0.4",
     "temp": "^0.8.3",
+    "url-regex": "^5.0.0",
     "winston": "^1.1.1"
   },
   "repository": {
