+++ src/smc-util/misc2.ts
@@ -8,8 +8,6 @@ it in a more modern ES 2018/Typescript/standard libraries approach.
 */
 
 const underscore = require("underscore");
-const urlRegex = require("url-regex");
-import { to_human_list } from "./misc";
 
 interface SplittedPath {
   head: string;
@@ -479,23 +477,4 @@ export function bind_methods(obj: any, method_names: string[]): void {
   for (let method_name of method_names) {
     obj[method_name] = obj[method_name].bind(obj);
   }
-}
-
-// used to test for URLs in a string
-const re_url = urlRegex({ exact: false, strict: false });
-
-// returns undefined if ok, otherwise an error message
-export function is_valid_username(str: string): string | undefined {
-  const name = str.toLowerCase();
-
-  const found = name.match(re_url);
-  if (found) {
-    return `URLs are not allowed. Found ${to_human_list(found)}`;
-  }
-
-  if (name.indexOf("mailto:") != -1 && name.indexOf("@") != -1) {
-    return "email addresses are not allowed";
-  }
-
-  return;
 }
