+++ src/smc-util/package.json
@@ -25,7 +25,6 @@
     "prop-types": "^15.5.10",
     "sha1": "^1.1.1",
     "underscore": "^1.9.1",
-    "url-regex": "^5.0.0",
     "uuid": "^3.0.1"
   },
   "repository": {
